refactor(ui): remove optional chaining warning; migrate Score import to domain; delete legacy model files
